+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [3.5.5]
 
 #### Added
 - - Added `IterableDecryptionFailureHandler` interface to handle decryption failures of PII information.